python-bindings: feature complete.
Old and new python code now lives in  python-pkg/. This is installed
in  regular locations by Makefile, but it is also a functional
pypi package which could be installed e. g., in a virtualenv.

The API contains pieces to send commands and receive data. See the
API docs and also the manual. There are also some proof-of-concepts
in doc/python-bindings.